integration-tests/requirements.txt
pytest
pytest-order
hypothesis
testcontainers
dbt-oracle
